#4: Add new speed commands to access project files and create project README file
The new functions `alb-org-project-files` and `alb-org-project-readme`
are bound to the speed keys `~` and `!` respectively.  These function
use the customisation in `alb-org-project-dirname-patterns` to transform
an Org-Mode heading into the filename of a project directory, and then
call `find-name-dired` on this filename.  This allows project files to
be quickly accessed from an Org-Mode heading

These changes also begin the use of the `alb-org` customisation group,
so that eventually the hard wired behaviour will be phased out of the
code